+++ fiddles/java/fiddle-0004-CSVToJsonEndpoint/springfox/boot-swagger/src/main/java/springfoxdemo/boot/swagger/web/Category.java
@@ -1,7 +1,0 @@
-package springfoxdemo.boot.swagger.web;
-
-public enum Category {
-    ONE,
-    TWO,
-    THREE
-}