+++ fiddles/java/fiddle-0004-CSVToJsonEndpoint/springfox/spring-java-swagger/src/main/java/springfoxdemo/java/swagger/SwaggerConfiguration.java
@@ -1,8 +1,0 @@
-package springfoxdemo.java.swagger;
-
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@EnableSwagger2
-public class SwaggerConfiguration {
-
-}