+++ fiddles/java/fiddle-0004-CSVToJsonEndpoint/springfox/spring-xml-swagger/src/main/java/springfoxdemo/xml/swagger/ApplicationSwaggerConfig.java
@@ -1,8 +1,0 @@
-package springfoxdemo.xml.swagger;
-
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@EnableSwagger2
-public class ApplicationSwaggerConfig {
-
-}