+++ fiddles/jquery/fiddle-0033-15Puzzle/app.js
@@ -1058,5 +1058,7 @@
             })
         }
     };
+
     document.body.addEventListener('DOMContentLoaded', app.controller.onDomContentLoaded(), false);
+
 })(window.app = window.app || {})