+++ fiddles/jquery/fiddle-0042-EniPuzzle/src/init.js
@@ -1,0 +1,12 @@
+app.controller = app.controller || {
+        onDomContentLoaded: function() {
+            var docElement = window.document.getElementById('fiddleHook');
+            app.docElement = new Board({
+                hook: docElement,
+                autoBind: true
+            })
+        }
+    };
+
+document.body.addEventListener('DOMContentLoaded', app.controller.onDomContentLoaded(), false);
+