+++ fiddles/nativeScript/fiddle-0011-SqliteNg2/puzzle/app/shared/config.js
@@ -5,7 +5,7 @@ var Config = (function () {
     return Config;
 }());
 Config.title = '15 Puzzle';
-Config.isDev = true;
+Config.isDev = false;
 Config.dbFile = 'highscore.db';
 Config.defaultLevel = 1;
 Config.transition = {
@@ -15,4 +15,3 @@ Config.transition = {
     }
 };
 exports.Config = Config;
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiY29uZmlnLmpzIiwic291cmNlUm9vdCI6IiIsInNvdXJjZXMiOlsiY29uZmlnLnRzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiI7QUFBQTtJQUFBO0lBV0EsQ0FBQztJQUFELGFBQUM7QUFBRCxDQUFDLEFBWEQ7QUFDUyxZQUFLLEdBQVcsV0FBVyxDQUFDO0FBQzVCLFlBQUssR0FBWSxJQUFJLENBQUM7QUFDdEIsYUFBTSxHQUFXLGNBQWMsQ0FBQztBQUNoQyxtQkFBWSxHQUFXLENBQUMsQ0FBQztBQUN6QixpQkFBVSxHQUFXO0lBQzFCLFVBQVUsRUFBRTtRQUNWLElBQUksRUFBRSxNQUFNO1FBQ1osS0FBSyxFQUFFLFFBQVE7S0FDaEI7Q0FDRixDQUFDO0FBVlMsd0JBQU0iLCJzb3VyY2VzQ29udGVudCI6WyJleHBvcnQgY2xhc3MgQ29uZmlnIHtcbiAgc3RhdGljIHRpdGxlOiBzdHJpbmcgPSAnMTUgUHV6emxlJztcbiAgc3RhdGljIGlzRGV2OiBib29sZWFuID0gdHJ1ZTtcbiAgc3RhdGljIGRiRmlsZTogc3RyaW5nID0gJ2hpZ2hzY29yZS5kYic7XG4gIHN0YXRpYyBkZWZhdWx0TGV2ZWw6IG51bWJlciA9IDE7XG4gIHN0YXRpYyB0cmFuc2l0aW9uOiBPYmplY3QgPSB7XG4gICAgdHJhbnNpdGlvbjoge1xuICAgICAgbmFtZTogXCJmbGlwXCIsXG4gICAgICBjdXJ2ZTogXCJsaW5lYXJcIlxuICAgIH1cbiAgfTtcbn1cbiJdfQ==