+++ fiddles/nativeScript/fiddle-0011-SqliteNg2/puzzle/app/shared/config.ts
@@ -1,6 +1,6 @@
 export class Config {
   static title: string = '15 Puzzle';
-  static isDev: boolean = true;
+  static isDev: boolean = false;
   static dbFile: string = 'highscore.db';
   static defaultLevel: number = 1;
   static transition: Object = {